Adding a functional timeline
Added the methods to retrieve all posts and to start listening to new posts.
The retrieve all methods posts will be called when we open the screen and the viewModel will get the result and sort it based on the timestamp value, so the most recent ones are the top at the list. When this step is done, we start listening for new posts, so we’ll display new posts to the user as soon as they’re available.