Added the Delete posts feature
Added the ability to remove your own posts. To do that, just swipe the message cell to left until the Delete button appears or until the end to automatically remove the post.
During testing an issues was found with .childAdded event type from Firebase, so I had to update it to .childChanged. There’s a comment on the code that better explains the issue.